Fix the markdown formatting in the table
To render correctly on GitHub it seems to need the pipe to be escaped otherwise it seems to see it as the end of the table. 

If you wanted to keep it copy/pasteable for plan text users looking at file in an editor it might be best to just remove it from the table. and switch to a bulleted list.